move net_utils.py to net.py
